requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.3
PyYAML==6.0
bandit==1.7.0
requests